# api/swagger.yaml
openapi: '3.0.2'
info:
  title: FoREST
  version: '1.0'
servers:
  - url: https://backend.mdma.haveachin.de/
  - url: http://localhost:8080

# ENDPOINTS ########################################################################
paths:
...
  /accounts/users:
    get:
      tags:
        - User-Accounts
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetUserAccount"
        500:
          description: Internal Server Error.

    post:
      tags:
        - User-Accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUserAccount"
      responses:
        201:
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserAccount"
        400:
          description: Bad Request.
        500:
          description: Internal Server Error.
  /accounts/users/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags:
        - User-Accounts
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserAccount"
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.

    put:
      tags:
        - User-Accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAccount"
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserAccount"
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.

    delete:
      tags:
        - User-Accounts
      responses:
        204:
          description: No Content.
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.
  /accounts/users/{id}/change-password:
    post:
      tags:
        - User-Accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChange"
      responses:
        204:
          description: No Content.
        400:
          description: Bad Request.
        401:
          description: Unauthorized.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.
...
components:
  schemas:
...
    ID:
      type: integer
      example: 1
...
    ResourceWithID:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PasswordChange:
      type: object
      properties:
        password:
          type: string
          example: password123
...
    UserAccount:
      type: object
      properties:
        username:
          type: string
          example: H4r4ldD3rH4ck3r
        roleId:
          type: integer
          example: 42
    GetUserAccount:
      allOf:
        - $ref: "#/components/schemas/ResourceWithID"
        - $ref: "#/components/schemas/UserAccount"
    PostUserAccount:
      allOf:
        - $ref: "#/components/schemas/ResourceWithID"
        - $ref: "#/components/schemas/UserAccount"
      type: object
      properties:
        password:
          type: string
          example: changeme
...
  parameters:
...
    ID:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"